disabled Objective-C support
(basically only support for the objcselector
keyword) and feature, so that the objc branch will be bootstrappable
using the 2.4.0 release. Not done in trunk to ease merging from the objc
branch. This commit must be undone first in case the objc stuff would ever
be merged to the fixes_2_4 branch (otherwise you'll get a ton of conflicts).

Manually created this patch.
  svn-id: @13779 or r13779